# server/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Proyecto Final Ecommerce Backend Nicolas Ippoliti'
paths:
  /users:
    post:
      summary: 'Register a new user'
      consumes:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'User object'
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        200:
          description: 'Successfully registered'
          schema:
            $ref: '#/definitions/User'
    put:
      summary: 'Update a user'
      consumes:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Updated user object'
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        200:
          description: 'Successfully updated'
          schema:
            $ref: '#/definitions/User'
  /users/{userId}:
    get:
      summary: 'Get a user'
      parameters:
        - in: 'path'
          name: 'userId'
          required: true
          type: 'string'
      responses:
        200:
          description: 'Successful retrieval of user'
          schema:
            $ref: '#/definitions/User'
    delete:
      summary: 'Delete a user'
      parameters:
        - in: 'path'
          name: 'userId'
          required: true
          type: 'string'
      responses:
        200:
          description: 'Successfully deleted user'
...
definitions:
  User:
    type: 'object'
    properties:
      id:
        type: 'string'
      name:
        type: 'string'
      email:
        type: 'string'
      role:
        type: 'string'
      cart:
        type: 'string'
    required:
      - name
      - email
      - password
      - role
      - cart
    unique:
      - email
...